Add broken WIP higher half kernel code

+++ kernel/kernel/mem/higher_half_utils.h
@@ -2,6 +2,6 @@
 
 #define KERNEL_OFFSET 0xC0000000
 #define KERNEL_START 0x10000
-#define V2P(a) ((a) - KERNEL_OFFSET)
-#define P2V(a) ((a) + KERNEL_OFFSET)
+#define V2P(a) (((uint32_t)(a)) - KERNEL_OFFSET)
+#define P2V(a) (((uint32_t)(a)) + KERNEL_OFFSET)
